requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.11.1
aiosignal==1.3.1
altair==5.5.0
annotated-types==0.7.0
anyio==3.7.1
APScheduler==3.10.1
attrs==25.3.0
azure-common==1.1.28
azure-core==1.32.0
azure-identity==1.19.0
azure-search==1.0.0b2
azure-search-documents==11.5.2
beautifulsoup4==4.13.3
black==25.1.0
blinker==1.9.0
cachetools==5.5.2
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
cryptography==43.0.3
distro==1.9.0
docutils==0.21.2
et_xmlfile==2.0.0
filetype==1.2.0
Flask==2.3.3
Flask-Cors==4.0.0
fpdf==1.7.2
frozenlist==1.5.0
gitdb==4.0.12
GitPython==3.1.44
gunicorn==21.2.0
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
idna==3.10
iniconfig==2.1.0
isodate==0.7.2
isort==6.0.1
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.7.1
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
Kivy==2.3.1
Kivy-Garden==0.1.5
lxml==5.3.2
markdown-it-py==3.0.0
markdown2==2.5.3
MarkupSafe==3.0.2
mdurl==0.1.2
msal==1.31.0
msal-extensions==1.2.0
msrest==0.7.1
multidict==6.1.0
mypy_extensions==1.1.0
narwhals==1.32.0
numpy==2.2.4
oauthlib==3.2.2
openai==1.99.2
openpyxl==3.1.5
packaging==24.2
pandas==2.2.3
pandas-stubs==2.2.3.241009
pathlib==1.0.1
pathspec==0.12.1
pdf2image==1.17.0
pillow==11.1.0
platformdirs==4.3.7
pluggy==1.5.0
portalocker==2.10.1
propcache==0.2.0
protobuf==5.29.4
psycopg2-binary==2.9.10
pyarrow==19.0.1
pycparser==2.22
pydantic==2.10.0
pydantic_core==2.27.0
pydeck==0.9.1
Pygments==2.19.1
PyJWT==2.9.0
PyPDF2==3.0.1
pytest==8.3.5
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-dotenv==1.0.0
python-json-logger==2.0.7
pytz==2025.2
referencing==0.36.2
requests==2.32.3
requests-oauthlib==2.0.0
rich==13.9.4
rpds-py==0.24.0
setuptools==80.9.0
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.6
streamlit==1.44.0
streamlit-feedback==0.1.3
tenacity==9.0.0
toml==0.10.2
tornado==6.4.2
tqdm==4.67.1
types-pytz==2024.2.0.20241003
typing_extensions==4.13.0
tzdata==2025.2
tzlocal==5.3.1
urllib3==2.3.0
watchdog==6.0.0
Werkzeug==3.1.3
yarl==1.17.1
langdetect>=1.0.9
deepeval>=1.3.0
scipy>=1.11.0
matplotlib>=3.8.0
# Azure storage
azure-storage-blob>=12.9.0
# locust for load testing
locust==2.42.6
# scikit-learn for cosine similarity check
scikit-learn==1.8.0
#SQL Alchemy for database interactions
SQLAlchemy==2.0.46
# pandas for data manipulation
pandas==2.2.3
